requirements.txt
absl-py==2.1.0
accelerate==0.25.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.4.3
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beartype==0.14.1
beautifulsoup4==4.12.3
braceexpand==0.1.7
cachetools==5.3.2
captum==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm @ file:///croot/comm_1709322850197/work
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
datasets==2.16.1
debugpy==1.8.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.7
-e git+https://github.com/neuroexplicit-saar/Discover-then-Name.git@03484baef77747596712cf971e402f77b9da5a9a#egg=dncbm
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
ExifRead-nocycle==3.0.1
fancy-einsum==0.0.3
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.36.0
imbalanced-learn==0.12.4
imblearn==0.0
img2dataset==1.45.0
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.20.0
jaxtyping==0.2.25
jedi @ file:///croot/jedi_1721058342488/work
Jinja2==3.1.2
joblib==1.3.2
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
lightning-utilities==0.10.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
ml-dtypes==0.2.0
mmcv==1.5.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.3.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.1
psutil==5.9.7
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
pytorchcv==0.0.67
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
sentry-sdk==1.39.1
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
soupsieve==2.5
-e git+https://github.com/neuroexplicit-saar/Discover-then-Name.git@03484baef77747596712cf971e402f77b9da5a9a#egg=sparse_autoencoder&subdirectory=sparse_autoencoder
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.9.20
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torch-summary==1.4.5
torchmetrics==1.3.0.post0
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformer-lens==1.12.0
transformers==4.36.2
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
